+++ src/entity/posts/Post.ts
@@ -1,0 +1,130 @@
+import {
+  Column,
+  Entity,
+  Index,
+  ManyToOne,
+  OneToMany,
+  PrimaryColumn,
+  TableInheritance,
+} from 'typeorm';
+import { Source } from '../Source';
+import { PostTag } from '../PostTag';
+import { PostKeyword } from '../PostKeyword';
+import { User } from '../User';
+
+@Entity()
+@TableInheritance({
+  column: { type: 'varchar', name: 'type', default: 'article' },
+})
+export class Post {
+  @PrimaryColumn({ type: 'text' })
+  id: string;
+
+  type: string;
+
+  @Column({ type: 'text' })
+  title: string;
+
+  @Column({ length: 14 })
+  @Index('IDX_post_shortid', { unique: true })
+  shortId: string;
+
+  @Column({ default: () => 'now()' })
+  @Index('IDX_post_createdAt', { synchronize: false })
+  createdAt: Date;
+
+  @Column({ default: () => 'now()' })
+  @Index('IDX_post_metadataChangedAt')
+  metadataChangedAt: Date;
+
+  @Column({ type: 'text', nullable: true })
+  @Index()
+  sourceId: string | null;
+
+  @ManyToOne(() => Source, (source) => source.posts, {
+    lazy: true,
+    onDelete: 'CASCADE',
+  })
+  source: Promise<Source>;
+
+  @Column({ default: false })
+  tweeted: boolean;
+
+  @Column({ default: 0 })
+  @Index('IDX_post_views')
+  views: number;
+
+  @Column({ type: 'integer', default: 0 })
+  @Index('IDX_post_score', { synchronize: false })
+  score: number;
+
+  @OneToMany(() => PostTag, (tag) => tag.post, { lazy: true })
+  tags: Promise<PostTag[]>;
+
+  @OneToMany(() => PostKeyword, (keyword) => keyword.post, { lazy: true })
+  keywords: Promise<PostKeyword[]>;
+
+  @Column({ type: 'text', nullable: true })
+  @Index('IDX_tags')
+  tagsStr: string;
+
+  @Column({ type: 'integer', default: 0 })
+  @Index('IDX_post_upvotes')
+  upvotes: number;
+
+  @Column({ type: 'integer', default: 0 })
+  @Index('IDX_post_comments')
+  comments: number;
+
+  @Column({ length: 36, nullable: true })
+  @Index('IDX_post_scout')
+  scoutId: string | null;
+
+  @ManyToOne(() => User, {
+    lazy: true,
+    onDelete: 'SET NULL',
+  })
+  scout: Promise<User>;
+
+  @Column({ length: 36, nullable: true })
+  @Index('IDX_post_author')
+  authorId: string | null;
+
+  @ManyToOne(() => User, {
+    lazy: true,
+    onDelete: 'SET NULL',
+  })
+  author: Promise<User>;
+
+  @Column({ default: true })
+  @Index('IDX_user_sentAnalyticsReport')
+  sentAnalyticsReport: boolean;
+
+  @Column({ default: 0 })
+  @Index('IDX_post_viewsThreshold')
+  viewsThreshold: number;
+
+  @Column({ nullable: true })
+  @Index('IDX_post_trending')
+  trending?: number;
+
+  @Column({ nullable: true })
+  @Index('IDX_post_last_trending')
+  lastTrending?: Date;
+
+  @Column({ type: 'integer', nullable: true })
+  @Index('IDX_post_discussion_score')
+  discussionScore?: number;
+
+  @Column({ default: false })
+  @Index('IDX_post_banned')
+  banned: boolean;
+
+  @Column({ default: false })
+  @Index('IDX_post_deleted')
+  deleted: boolean;
+
+  @Column({ nullable: true, type: 'tsvector', select: false })
+  @Index('IDX_post_tsv')
+  tsv: unknown;
+}
