+++ src/entity/posts/index.ts
@@ -1,0 +1,4 @@
+export * from './Post';
+export * from './ArticlePost';
+export * from './SharePost';
+export * from './utils';
