+++ src/entity/posts/utils.ts
@@ -1,204 +1,16 @@
-import {
-  ChildEntity,
-  Column,
-  DataSource,
-  Entity,
-  EntityManager,
-  In,
-  Index,
-  ManyToOne,
-  OneToMany,
-  OneToOne,
-  PrimaryColumn,
-  TableInheritance,
-} from 'typeorm';
+import { DataSource, EntityManager, In } from 'typeorm';
+import { SubmissionFailErrorKeys } from '../../errors';
 import shortid from 'shortid';
 import * as he from 'he';
-import { PostTag } from './PostTag';
-import { Source } from './Source';
-import { User } from './User';
-import { PostKeyword } from './PostKeyword';
-import { Keyword } from './Keyword';
-import { uniqueifyArray } from '../common';
-import { validateAndApproveSubmission } from './Submission';
-import { SubmissionFailErrorKeys } from '../errors';
+import { Keyword } from '../Keyword';
+import { uniqueifyArray } from '../../common';
+import { User } from '../User';
 import { FastifyLoggerInstance } from 'fastify';
-
-export type TocItem = { text: string; id?: string; children?: TocItem[] };
-export type Toc = TocItem[];
-
-@Entity()
-@TableInheritance({
-  column: { type: 'varchar', name: 'type', default: 'article' },
-})
-export class Post {
-  @PrimaryColumn({ type: 'text' })
-  id: string;
-
-  type: string;
-
-  @Column({ type: 'text' })
-  title: string;
-
-  @Column({ length: 14 })
-  @Index('IDX_post_shortid', { unique: true })
-  shortId: string;
-
-  @Column({ default: () => 'now()' })
-  @Index('IDX_post_createdAt', { synchronize: false })
-  createdAt: Date;
-
-  @Column({ default: () => 'now()' })
-  @Index('IDX_post_metadataChangedAt')
-  metadataChangedAt: Date;
-
-  @Column({ type: 'text', nullable: true })
-  @Index()
-  sourceId: string | null;
-
-  @ManyToOne(() => Source, (source) => source.posts, {
-    lazy: true,
-    onDelete: 'CASCADE',
-  })
-  source: Promise<Source>;
-
-  @Column({ default: false })
-  tweeted: boolean;
-
-  @Column({ default: 0 })
-  @Index('IDX_post_views')
-  views: number;
-
-  @Column({ type: 'integer', default: 0 })
-  @Index('IDX_post_score', { synchronize: false })
-  score: number;
-
-  @OneToMany(() => PostTag, (tag) => tag.post, { lazy: true })
-  tags: Promise<PostTag[]>;
-
-  @OneToMany(() => PostKeyword, (keyword) => keyword.post, { lazy: true })
-  keywords: Promise<PostKeyword[]>;
-
-  @Column({ type: 'text', nullable: true })
-  @Index('IDX_tags')
-  tagsStr: string;
-
-  @Column({ type: 'integer', default: 0 })
-  @Index('IDX_post_upvotes')
-  upvotes: number;
-
-  @Column({ type: 'integer', default: 0 })
-  @Index('IDX_post_comments')
-  comments: number;
-
-  @Column({ length: 36, nullable: true })
-  @Index('IDX_post_scout')
-  scoutId: string | null;
-
-  @ManyToOne(() => User, {
-    lazy: true,
-    onDelete: 'SET NULL',
-  })
-  scout: Promise<User>;
-
-  @Column({ length: 36, nullable: true })
-  @Index('IDX_post_author')
-  authorId: string | null;
-
-  @ManyToOne(() => User, {
-    lazy: true,
-    onDelete: 'SET NULL',
-  })
-  author: Promise<User>;
-
-  @Column({ default: true })
-  @Index('IDX_user_sentAnalyticsReport')
-  sentAnalyticsReport: boolean;
-
-  @Column({ default: 0 })
-  @Index('IDX_post_viewsThreshold')
-  viewsThreshold: number;
-
-  @Column({ nullable: true })
-  @Index('IDX_post_trending')
-  trending?: number;
-
-  @Column({ nullable: true })
-  @Index('IDX_post_last_trending')
-  lastTrending?: Date;
-
-  @Column({ type: 'integer', nullable: true })
-  @Index('IDX_post_discussion_score')
-  discussionScore?: number;
-
-  @Column({ default: false })
-  @Index('IDX_post_banned')
-  banned: boolean;
-
-  @Column({ default: false })
-  @Index('IDX_post_deleted')
-  deleted: boolean;
-
-  @Column({ nullable: true, type: 'tsvector', select: false })
-  @Index('IDX_post_tsv')
-  tsv: unknown;
-}
-
-@ChildEntity('article')
-export class ArticlePost extends Post {
-  @Column({ nullable: true })
-  publishedAt?: Date;
-
-  @Column({ type: 'text' })
-  @Index({ unique: true })
-  url: string;
-
-  @Column({ type: 'text', nullable: true })
-  @Index({ unique: true })
-  canonicalUrl?: string;
-
-  @Column({ type: 'text', nullable: true })
-  image?: string;
-
-  @Column({ type: 'float', nullable: true })
-  ratio?: number;
-
-  @Column({ type: 'text', nullable: true })
-  placeholder?: string;
-
-  @Column({ type: 'text', nullable: true })
-  siteTwitter?: string;
-
-  @Column({ type: 'text', nullable: true })
-  creatorTwitter?: string;
-
-  @Column({ nullable: true })
-  readTime?: number;
-
-  @Column({ type: 'text', nullable: true })
-  description?: string;
-
-  @Column({ type: 'jsonb', nullable: true })
-  toc?: Toc;
-
-  @Column({ type: 'text', nullable: true })
-  summary?: string;
-}
-
-@ChildEntity('share')
-export class SharePost extends Post {
-  @Column({ type: 'text' })
-  sharedPostId: string;
-
-  @OneToOne(() => Post, { lazy: true, onDelete: 'SET NULL' })
-  sharedPost: Promise<Post>;
-}
-
-export interface SearchPostsResult {
-  id: string;
-  title: string;
-  highlight?: string;
-}
+import { PostTag } from '../PostTag';
+import { PostKeyword } from '../PostKeyword';
+import { validateAndApproveSubmission } from '../Submission';
+import { ArticlePost, Toc } from './ArticlePost';
+import { Post } from './Post';
 
 export type PostStats = {
   numPosts: number;
