+++ src/schema/bookmarks.ts
@@ -121,6 +121,11 @@ export const typeDefs = /* GraphQL */ `
       Id of the list to retrieve bookmarks from
       """
       listId: ID
+
+      """
+      Array of supported post types
+      """
+      supportedTypes: [String!]
     ): PostConnection! @auth
 
     """
@@ -171,6 +176,11 @@ export const typeDefs = /* GraphQL */ `
       Id of the list to retrieve bookmarks from
       """
       listId: ID
+
+      """
+      Array of supported post types
+      """
+      supportedTypes: [String!]
     ): PostConnection! @auth
   }
 `;
@@ -179,6 +189,7 @@ interface BookmarksArgs extends ConnectionArguments {
   now: Date;
   unreadOnly: boolean;
   listId: string;
+  supportedTypes?: string[];
 }
 
 interface BookmarkPage extends Page {
@@ -344,13 +355,16 @@ export const resolvers: IResolvers<any, Context> = traceResolvers({
     ) => {
       const hits: { title: string }[] = await ctx.con.query(
         `
-        WITH search AS (${getSearchQuery('$2')})
-        select ts_headline(process_text(title), search.query,
-        'StartSel = <strong>, StopSel = </strong>') as title
-        from post INNER JOIN bookmark ON bookmark."postId" = post.id AND bookmark."userId" = $1, search
-        where tsv @@ search.query
-        order by views desc
-        limit 5;
+          WITH search AS (${getSearchQuery('$2')})
+          select ts_headline(process_text(title), search.query,
+                             'StartSel = <strong>, StopSel = </strong>') as title
+          from post
+                 INNER JOIN bookmark ON bookmark."postId" = post.id AND
+                                        bookmark."userId" = $1,
+               search
+          where tsv @@ search.query
+          order by views desc
+            limit 5;
         `,
         [ctx.userId, query],
       );
